Update login_page.html
Done the changes as you have mentioned.
Did the indentation part for contents of <ul>
Changed the id and for attribute of both username and password as mentioned.
Changed button type to submit 
Have put required attribute in all inputs
Added the link part in href and have made the whole part as link. Do check the link part because I don't know whether what I have done is what you wanted me to change.
If any more changes have to be made in the code, then let me know.